+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # [IN PROGRESS]
 
 * Hide 'Item details' in Action menu when item is virtual. Refs. UICHKOUT-871.
+* Hide Patron and Staff Info in Action menu when DCB patron. Refs. UICHKOUT-872.
 
 ## [10.0.1](https://github.com/folio-org/ui-checkout/tree/v10.0.1) (2023-10-23)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v10.0.0...v10.0.1)